Add files via upload
Browser only under stand DOM
DOM: 
The Document Object Model (DOM) is a programming interface that represents the structure of a web document in memory. It allows programs to change the document's style, content, and structure

Web browsers understand HTML because it's a text-based language that tells browsers how to display the content of a web page: 
 
HTML tells the browser how to display content
HTML (Hypertext Markup Language) is the backbone of website design and function. It tells the browser how to display text, images, and other multimedia on a web pag
 
HTML TAGS
Creating html file using .html extention

HTML starts with "<>" and ends with "</>"

<html>
	<head></head>
	<body>
		this body content will be display to the end user
	</body>
</html>
 
 Revise - DOM
 <html>
 <head>
  <title></title>
 </head>
 <body></body>
 </html>
 
 docker run --detach --name mydb --env MARIADB_ROOT_PASSWORD=root -p 3306:3306  mariadb:latest